In payment create form, incorrect choice in field "Recipient of payment" and "payer"
Initially it was misnamed Association <association name="payments:payer"> and <association name="payments:beneficiary"> in paymentsModel.xml. Due to the inability to fix, it was decided to change their name in the localization files.
That solved the problem graphically. But in other places, such as invoice templates, had to rearrange the payer and the recipient. Perhaps in other templates incorrectly filled in these fields.